requirements.txt
Pillow~=10.0.0
requests~=2.31.0
transformers~=4.31.0
pydantic~=2.1.1
python-decouple~=3.8
Google-Images-Search~=1.4.6
datamodel-code-generator~=0.21.4
replicate~=0.11.0
openai~=0.27.8
g4f~=0.0.1.9
yadisk~=1.3.3
google-api-python-client~=2.97.0
google-auth-oauthlib~=1.0.0
oauthlib~=3.2.2
requests-oauthlib~=1.3.1
flask~=2.3.2
genson~=1.2.2
jsonschema~=4.17.3